+++ src/@lib/equalities/shallowEquals.ts
@@ -1,5 +1,8 @@
 // Record 타입: 키-값 쌍을 가진 객체 타입을 정의할 때 사용
-export function shallowEquals<T>(objA: T, objB: T): boolean {
+export function shallowEquals<T extends Record<string, unknown>>(
+  objA: T,
+  objB: T,
+): boolean {
   if (objA === objB) {
     // 같은 객체 참조 확인
     return true;
@@ -24,15 +27,9 @@ export function shallowEquals<T>(objA: T, objB: T): boolean {
   }
 
   // 속성 값 비교
-  for (const key of keysA) {
-    // if (!objB.hasOwnProperty(key) || objA[key] !== objB[key]) {
-    if (
-      !Object.prototype.hasOwnProperty.call(objB, key) ||
-      objA[key] !== objB[key]
-    ) {
-      return false;
-    }
-  }
-
-  return true;
+  return keysA.every(
+    (key) =>
+      Object.prototype.hasOwnProperty.call(objB, key) &&
+      objA[key] === objB[key],
+  );
 }
