+++ src/App.tsx
@@ -1,23 +1,12 @@
 import React from "react";
-import {
-  ComplexForm,
-  ItemList,
-} from "./components";
 import { AppProviders } from "./providers";
-import { Layout } from "./components/Layout";
+import { ComplexForm, ItemList } from "./features";
+import { Layout } from "./shared/components/layout/Layout";
 
 // 메인 App 컴포넌트
 const App: React.FC = () => {
-  // Theme Provider로 둘러싸져있어야 훅을 호출할 수 있음...!!-> 지금 외부에서 호출중 ..
-  // const { theme } = useTheme();
-
   return (
     <AppProviders>
-      {/* <div
-        className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
-      > */}
-      {/* <Header /> */}
-      {/* <div className="container mx-auto px-4 py-8"> */}
       <Layout>
         <div className="flex flex-col md:flex-row">
           <div className="w-full md:w-1/2 md:pr-4">
@@ -28,9 +17,6 @@ const App: React.FC = () => {
           </div>
         </div>
       </Layout>
-      {/* </div> */}
-      {/* <NotificationSystem /> */}
-      {/* </div> */}
     </AppProviders>
   );
 };
