--- src/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./Header";
-export * from "./ItemList";
-export * from "./ComplexForm";
-export * from "./NotificationSystem";
