--- src/contexts/ThemeContext.tsx
@@ -1,16 +1,0 @@
-import { createContext, useContext } from "react";
-
-type ThemeContextProps = {
-  theme: string;
-  toggleTheme: () => void;
-};
-
-export const ThemeContext = createContext<ThemeContextProps | null>(null);
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (!context) {
-    throw new Error("useTheme must be used within an ThemeProvider");
-  }
-  return context;
-};
