--- src/contexts/index.ts
@@ -1,3 +1,0 @@
-export * from "./ThemeContext";
-export * from "./UserContext";
-export * from "./NotificationContext";
