+++ src/core/notifications/contexts/NotificationContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+import { NotificationContextProps } from "../types";
+
+export const NotificationContext =
+  createContext<NotificationContextProps | null>(null);
