+++ src/core/theme/contexts/ThemeContext.tsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { ThemeContextProps } from "../types";
+
+export const ThemeContext = createContext<ThemeContextProps | null>(null);
