+++ src/core/theme/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useTheme";
