+++ src/core/theme/types/index.ts
@@ -1,0 +1,4 @@
+export interface ThemeContextProps {
+  theme: string;
+  toggleTheme: () => void;
+}
