+++ src/core/user/contexts/UserContext.tsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { UserContextProps } from "../types";
+
+export const UserContext = createContext<UserContextProps | null>(null);
