+++ src/core/user/providers/UserProvider.tsx
@@ -1,21 +1,20 @@
 import { ReactNode, useCallback, useMemo, useState } from "react";
-import { useNotification, UserContext } from "../contexts";
-
-type User = {
-  id: number;
-  name: string;
-  email: string;
-};
+import { UserContext } from "../contexts";
+import { useNotification } from "../../notifications";
+import { User } from "../types";
 
 export const UserProvider = ({ children }: { children: ReactNode }) => {
   const [user, setUser] = useState<User | null>(null);
   const { addNotification } = useNotification();
 
   // context 같이 써도 되나? 함수 단일 원칙에 위배되나 결합이 느슨해서 괜찮...??
-  const login = useCallback((email: string) => {
-    setUser({ id: 1, name: "홍길동", email });
-    addNotification("성공적으로 로그인되었습니다.", "success");
-  }, [addNotification]);
+  const login = useCallback(
+    (email: string) => {
+      setUser({ id: 1, name: "홍길동", email });
+      addNotification("성공적으로 로그인되었습니다.", "success");
+    },
+    [addNotification],
+  );
 
   const logout = useCallback(() => {
     setUser(null);
