+++ src/features/form/components/ComplexForm.tsx
@@ -1,6 +1,6 @@
 import { memo, useCallback, useState } from "react";
-import { renderLog } from "../utils";
-import { useNotification } from "../contexts";
+import { renderLog } from "../../../utils";
+import { useNotification } from "../../../core";
 
 // ComplexForm 컴포넌트
 export const ComplexForm: React.FC = memo(() => {
@@ -13,27 +13,36 @@ export const ComplexForm: React.FC = memo(() => {
     preferences: [] as string[],
   });
 
-  const handleSubmit = useCallback((e: React.FormEvent) => {
-    e.preventDefault();
-    addNotification("폼이 성공적으로 제출되었습니다", "success");
-  }, [addNotification]);
+  const handleSubmit = useCallback(
+    (e: React.FormEvent) => {
+      e.preventDefault();
+      addNotification("폼이 성공적으로 제출되었습니다", "success");
+    },
+    [addNotification],
+  );
 
-  const handleInputChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: name === "age" ? parseInt(value) || 0 : value,
-    }));
-  }, [setFormData]);
+  const handleInputChange = useCallback(
+    (e: React.ChangeEvent<HTMLInputElement>) => {
+      const { name, value } = e.target;
+      setFormData((prev) => ({
+        ...prev,
+        [name]: name === "age" ? parseInt(value) || 0 : value,
+      }));
+    },
+    [setFormData],
+  );
 
-  const handlePreferenceChange = useCallback((preference: string) => {
-    setFormData((prev) => ({
-      ...prev,
-      preferences: prev.preferences.includes(preference)
-        ? prev.preferences.filter((p) => p !== preference)
-        : [...prev.preferences, preference],
-    }));
-  }, [setFormData]);
+  const handlePreferenceChange = useCallback(
+    (preference: string) => {
+      setFormData((prev) => ({
+        ...prev,
+        preferences: prev.preferences.includes(preference)
+          ? prev.preferences.filter((p) => p !== preference)
+          : [...prev.preferences, preference],
+      }));
+    },
+    [setFormData],
+  );
 
   return (
     <div className="mt-8">
