+++ src/features/items/components/ItemList.tsx
@@ -1,31 +1,14 @@
-import { memo, useCallback, useMemo, useState } from "react";
-import { generateItems, renderLog } from "../utils";
-import { useTheme } from "../contexts";
-
-// 타입 정의
-// interface Item {
-//   id: number;
-//   name: string;
-//   category: string;
-//   price: number;
-// }
+import { memo, useState } from "react";
+import { renderLog } from "../../../utils";
+import { useItems } from "../hooks";
+import { useTheme } from "../../../core";
 
 // ItemList 컴포넌트
 export const ItemList: React.FC = memo(() => {
   renderLog("ItemList rendered");
-  const initialItems = useMemo(() => generateItems(1000), []);
-  const [items, setItems] = useState(initialItems);
+  const { items, addItems } = useItems();
   const [filter, setFilter] = useState("");
   const { theme } = useTheme();
-
-  // 관심사 분리하기 위해서 옮겨옴
-  // 이거 useCallback해도 안되는 이유는 위에 useTheme 훅을 내부에서 쓰고있어서 컴포넌트가 다시 렌더링 되버림 
-  const addItems = useCallback(() => {
-    setItems((prevItems) => [
-      ...prevItems,
-      ...generateItems(1000, prevItems.length),
-    ]);
-  }, [setItems]);
 
   const filteredItems = items.filter(
     (item) =>
